# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Picroom API
  version: 1.0.0
  description: |
    Picroom — self-hosted image hosting service.

    Two API surfaces:
    1. REST/JSON under `/api/v1/` — first-class management API.
    2. AWS S3-compatible under `/s3/` — drop-in replacement for object storage
       clients (PicGo, rclone, AWS CLI).

    All endpoints require authentication except where noted.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Picroom Maintainers
    url: https://github.com/picroom/picroom
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://{host}
    description: Self-hosted production
    variables:
      host:
        default: picroom.example.com
...
paths:
...
  /api/v1/images/{imageId}/acl:
    parameters:
      - $ref: '#/components/parameters/ImageId'
    get:
      tags: [images]
      summary: List the image's ACL grants
      description: >
        Owner or `Image/Update` (manager/admin) only. Returns the resource's
        grant set; evaluation order is explicit deny, ownership, team role,
        ACL allow, global role, default deny (spec §10.3).
      operationId: listImageAcl
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AclGrant'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [images]
      summary: Replace the image's full ACL grant set
      description: Replace semantics — the request body is the complete grant list; idempotent on repeat.
      operationId: replaceImageAcl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [grants]
              properties:
                grants:
                  type: array
                  items:
                    $ref: '#/components/schemas/AclGrant'
      responses:
        '200':
          description: The stored grant set
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AclGrant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/images/{imageId}/acl/{subjectType}/{subjectId}:
    parameters:
      - $ref: '#/components/parameters/ImageId'
      - name: subjectType
        in: path
        required: true
        schema:
          type: string
          enum: [user, team]
      - name: subjectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [images]
      summary: Remove every grant for one subject on the image
      operationId: revokeImageAcl
      responses:
        '204':
          description: Removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: picroom_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sigv4:
      type: apiKey
      in: header
      name: Authorization
      description: AWS Signature Version 4
  parameters:
    ImageId:
      in: path
      name: imageId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples: [quota_exceeded, validation_failed]
        message:
          type: string
          description: Human-readable message
        request_id:
          type: string
          format: uuid
        details:
          type: object
          additionalProperties: true
...
security:
  - cookieAuth: []
  - bearerAuth: []